Fixes: US030 Added Backend processing module for creating video chunks  (#24)
## Summary

This PR adds a backend module to the src, which lays the ground work for
processing the mcap file to a format that can be efficiently send over
wire to the UI. Here, we follow a "pipeline of executors" approach. That
is, for each topic, we generate a sequence of executors that based on
its schema type. When the first executor is called, it is responsible
for getting the data from its next executors and processing it and
finally populating a buffer pool.

The webserver later would just query the buffer pool and send the
requested chunks to the UI.

This PR makes it easy to add support for Lidar, or GPS or any other
schema type. It also make it easy to add caching, parallelization and
any other optimizations.

Currently the memory profile of processing a single topic is as shown in
the following screenshot. This is good for the version 1, but would be
optimized later.

<img width="1338" height="465" alt="Screenshot 2025-10-13 at 1 52 05 AM"
src="https://github.com/user-attachments/assets/cdda00bc-4a94-4e4c-86cd-144d97c9d7c3"
/>


## 🔗 Issue

Fixes US-030

## Checklist (must be completed before merge)

- [x] PR title follows format: `Fixes #<issue_number> -
(chore/fix/bug/feat/tests) short description`
- [x] PR description includes the issue being fixed, how it fixes it, if
it breaks anything, and tests added
- [x] Code is formatted with **Black** (`black .`)
- [x] Code passes **Flake8** linting (`flake8`)
- [ ] Types check cleanly with **mypy** (`mypy .`)
- [x] Commit history is clean, relevant, and logically grouped
- [x] No duplicated logic, code is well-designed and maintainable
- [x] Code coverage ≥ 90% on new code and overall coverage is not
reduced
- [x] CI pipeline passes (tests + checks)
- [ ] At least one reviewer has approved with meaningful feedback (not
just “LGTM”)